Move SuluAutomationBundle services into Infrastructure namespace
