+++ Resources/config/automation.xml
@@ -4,7 +4,7 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
         <service id="sulu_content.content.automation.publish_handler"
-                 class="Sulu\Bundle\ContentBundle\Content\Application\Automation\Handler\ContentEntityPublishHandler">
+                 class="ContentPublishTaskHandler">
             <argument type="service" id="sulu_content.content_manager"/>
             <argument type="service" id="doctrine.orm.entity_manager"/>
             <argument type="service" id="translator"/>
@@ -13,7 +13,7 @@
         </service>
 
         <service id="sulu_content.content.automation.unpublish_handler"
-                 class="Sulu\Bundle\ContentBundle\Content\Application\Automation\Handler\ContentEntityUnpublishHandler">
+                 class="ContentUnpublishTaskHandler">
             <argument type="service" id="sulu_content.content_manager"/>
             <argument type="service" id="doctrine.orm.entity_manager"/>
             <argument type="service" id="translator"/>
